First stab at the MainViewModel
It's a ReactiveObject.  The LoadGpxDataCommand invokes the fileChooser ui (injected in the constructor), reads the returned stream and replays the points - updating the CurrentPoint property as the points replay.

There's no error handling yet - it'll crash if the stream returned by the fileChooser doesn't contain valid GPX data.
And it should probably do something different when the Replay observable completes.