Using an ObservableAsPropertyHelper instead of the currentPoint backing field
This feels like a more "ReactiveUI-y" way of doing it ...and it I means I don't have to worry about the scheduler (by default, the OAPH dispatches to the main thread).

Also, bring the OAPH outside of the command action ...RegisterAsyncFunction gives me an observable of observables. A new inner observable is posted every time the command executes - so I can just switch on the inner observables to allow loading a new file while one is running.