Bugfix: Replay waits for the time of the last section before completing
The Replay observable would produce the last GpxSection and then complete immediately. It should wait for the time interval between the start and end points in the last section before completing.

The simplest way to achieve this was to stick a null GpxSection at the end of the enumerable. The observable will then...

+ produce the last section
+ wait for the time interval between the start and end points in the last section
+ produce the null section
+ complete immediately